[cargo] Simplify build.rs with helper crate
