+++ rxx/Cargo.toml
@@ -14,6 +14,7 @@
 
 [workspace]
 members = [
+  "cxx-buildgen",
   "macros",
   "namegen",
   "rslogger",
@@ -38,6 +39,7 @@ named-lock = "^0.4.0"
 macros = { path = "macros" }
 namegen = { path = "namegen" }
 rslogger = { path = "rslogger" }
+cxx-buildgen = { path = "cxx-buildgen" }
 
 # This is debug
 [profile.dev]
